feat(mobile): create qr code scanner package
