+++ suite-native/module-accounts-import/package.json
@@ -31,6 +31,7 @@
         "@suite-native/module-devices": "workspace:*",
         "@suite-native/module-settings": "workspace:*",
         "@suite-native/navigation": "workspace:*",
+        "@suite-native/qr-code-scanner": "workspace:*",
         "@trezor/connect": "workspace:9.0.4",
         "@trezor/icons": "workspace:*",
         "@trezor/styles": "workspace:*",
