+++ suite-native/module-accounts-import/tsconfig.json
@@ -33,6 +33,7 @@
         { "path": "../module-devices" },
         { "path": "../module-settings" },
         { "path": "../navigation" },
+        { "path": "../qr-code-scanner" },
         { "path": "../../packages/connect" },
         { "path": "../../packages/icons" },
         { "path": "../../packages/styles" },
