+++ suite-native/qr-code-scanner/package.json
@@ -1,0 +1,20 @@
+{
+    "name": "@suite-native/qr-code-scanner",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
+        "type-check": "tsc --build"
+    },
+    "dependencies": {
+        "@react-navigation/core": "^6.2.2",
+        "@suite-native/atoms": "workspace:*",
+        "@trezor/styles": "workspace:*",
+        "expo-barcode-scanner": "^12.1.0",
+        "react": "18.2.0",
+        "react-native": "0.70.5"
+    }
+}
