+++ suite-native/qr-code-scanner/src/components/CameraPermissionError.tsx
@@ -26,13 +26,14 @@ export const CameraPermissionError = ({
 
     return (
         <Box style={applyStyle(permissionTextContainerStyle)}>
-            <Text align="center">Camera access denied.</Text>
+            <Text align="center">Camera access denied. {permissionStatus}</Text>
             {permissionStatus === PermissionStatus.DENIED && (
                 <Text align="center">Please allow camera access in your device settings.</Text>
             )}
             {permissionStatus === PermissionStatus.UNDETERMINED && (
                 <Text align="center">Please allow camera access to your camera. </Text>
             )}
+
             <Button onPress={onPermissionRequest} style={applyStyle(grantPermissionButtonStyle)}>
                 Grant permission
             </Button>
